adapt review comments
- rename `Action` to `Command` for Handler method routing object
- rename EventsApiAction to EventsApiCommand
- rename InteractiveAction to InteractiveCommand
